Implemented approach CommonJS
